requirements.txt
# web server
Flask==0.12.2

# frameworks
Flask-Admin
Flask-BabelEx

# config
Flask-Env

# database
Flask-SQLAlchemy
Flask-Migrate
psycopg2

# security
auth0-python==3.0.0
python-jose==1.3.2
Flask-Login==0.4.0